Move port attribute provider into kernel
