+++ src/kernels/port/portAttributeProvider.node.ts
@@ -4,11 +4,11 @@
 import { inject, injectable } from 'inversify';
 import { workspace } from 'vscode';
 import { CancellationToken, PortAttributes, PortAttributesProvider, PortAutoForwardAction } from 'vscode';
-import { NotebookStarter } from '../../../kernels/jupyter/launcher/notebookStarter.node';
-import { KernelLauncher } from '../../../kernels/raw/launcher/kernelLauncher.node';
-import { IExtensionSyncActivationService } from '../../activation/types';
-import { traceError } from '../../logging';
-import { IDisposableRegistry } from '../types';
+import { NotebookStarter } from '../jupyter/launcher/notebookStarter.node';
+import { KernelLauncher } from '../raw/launcher/kernelLauncher.node';
+import { IExtensionSyncActivationService } from '../../platform/activation/types';
+import { traceError } from '../../platform/logging';
+import { IDisposableRegistry } from '../../platform/common/types';
 
 @injectable()
 export class PortAttributesProviders implements PortAttributesProvider, IExtensionSyncActivationService {
