+++ src/kernels/serviceRegistry.node.ts
@@ -42,10 +42,14 @@ import { ServerConnectionType } from './jupyter/launcher/serverConnectionType';
 import { CellOutputDisplayIdTracker } from './execution/cellDisplayIdTracker';
 import { DebugLocationTrackerFactory } from './debugger/debugLocationTrackerFactory';
 import { Activation } from './activation.node';
+import { PortAttributesProviders } from './port/portAttributeProvider.node';
 
 export function registerTypes(serviceManager: IServiceManager, isDevMode: boolean) {
     serviceManager.addSingleton<IExtensionSingleActivationService>(IExtensionSingleActivationService, Activation);
-
+    serviceManager.addSingleton<IExtensionSyncActivationService>(
+        IExtensionSyncActivationService,
+        PortAttributesProviders
+    );
     serviceManager.addSingleton<IRawNotebookSupportedService>(
         IRawNotebookSupportedService,
         RawNotebookSupportedService
